+++ sdk/core/azure-core/azure/core/credentials.py
@@ -4,8 +4,10 @@
 # license information.
 # -------------------------------------------------------------------------
 from collections import namedtuple
-from typing import Any, NamedTuple, Optional
+from typing import Any, NamedTuple, Optional, Union
 from typing_extensions import Protocol, runtime_checkable
+import time
+import abc
 
 
 class AccessToken(NamedTuple):
@@ -52,6 +54,7 @@ __all__ = [
     "AccessToken",
     "AzureNamedKeyCredential",
     "TokenCredential",
+    "StaticTokenCredential",
 ]
 
 
@@ -163,3 +166,26 @@ class AzureNamedKeyCredential:
         if not isinstance(name, str) or not isinstance(key, str):
             raise TypeError("Both name and key must be strings.")
         self._credential = AzureNamedKey(name, key)
+
+
+class StaticTokenCredential(abc.ABC):
+    """Authenticates with a previously acquired access token
+    Note that an access token is valid only for certain resources and eventually expires. This credential is therefore
+    quite limited. An application using it must ensure the token is valid and contains all claims required by any
+    service client given an instance of this credential.
+
+    :param access_token: The pre-acquired access token.
+    :type access_token: str or ~azure.core.credentials.AccessToken
+    :keyword int expire_in: The number of seconds the token is valid. Defaults to 86400 seconds (1 day).
+        Only used if access_token is a str.
+    """
+    def __init__(self, access_token: Union[str, AccessToken], *, expire_in: int = 86400) -> None:
+        if isinstance(access_token, AccessToken):
+            self._token = access_token
+        else:
+            self._token = AccessToken(token=access_token, expires_on=time.time()+expire_in)
+
+    def get_token(self, *scopes: str, **kwargs: Any) -> AccessToken:    # pylint:disable=unused-argument
+        """get_token is the only method a credential must implement"""
+
+        return self._token