+++ sdk/core/azure-core/azure/core/credentials_async.py
@@ -2,9 +2,10 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
-from typing import Any, Optional
+from typing import Any, Optional, Union
 from typing_extensions import Protocol, runtime_checkable
 from .credentials import AccessToken as _AccessToken
+import time
 
 
 @runtime_checkable
@@ -38,3 +39,35 @@ class AsyncTokenCredential(Protocol):
 
     async def __aexit__(self, exc_type, exc_value, traceback) -> None:
         pass
+
+
+class AsyncStaticTokenCredential:
+    """Authenticates with a previously acquired access token
+    Note that an access token is valid only for certain resources and eventually expires. This credential is therefore
+    quite limited. An application using it must ensure the token is valid and contains all claims required by any
+    service client given an instance of this credential.
+
+    :param access_token: The pre-acquired access token.
+    :type access_token: str or ~azure.core.credentials.AccessToken
+    :keyword int expire_in: The number of seconds the token is valid. Defaults to 86400 seconds (1 day).
+        Only used if access_token is a str.
+    """
+    def __init__(self, access_token: Union[str, _AccessToken], *, expire_in: int = 86400) -> None:
+        if isinstance(access_token, _AccessToken):
+            self._token = access_token
+        else:
+            self._token = _AccessToken(token=access_token, expires_on=time.time()+expire_in)
+
+    async def close(self):
+        pass
+
+    async def __aenter__(self):
+        return self
+
+    async def __aexit__(self, *args):
+        await self.close()
+
+    async def get_token(self, *scopes: str, **kwargs: Any) -> _AccessToken:    # pylint:disable=unused-argument
+        """get_token is the only method a credential must implement"""
+
+        return self._token