+++ sdk/core/azure-core/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- Added `StaticTokenCredential` and `AsyncStaticTokenCredential` support.
+
 ### Breaking Changes
 
 ### Bugs Fixed